formularios

+++ src/app/library/libros/libros.component.ts
@@ -33,14 +33,16 @@ export class LibrosComponent implements OnInit {
   }
 
   deleteLibro(libro: Libro){
-    this.libroService.delete(libro.isbn).subscribe({
-      next: (response: any) => {
-        console.log('Libro eliminado');
-        this.loadLibros();
-      },
-      error: (error: any) => {
-        console.log('Error al eliminar libro', error);
-      }
-    });
+    if(libro.isbn != null){
+      this.libroService.delete(libro.isbn).subscribe({
+        next: (response: any) => {
+          console.log('Libro eliminado');
+          this.loadLibros();
+        },
+        error: (error: any) => {
+          console.log('Error al eliminar libro', error);
+        }
+      });
+    }
   }
 }
